requirements.txt
flask==3.0.2
numpy==1.26.4
pandas==2.2.1
openai==1.12.0
python-dotenv==1.0.1
scikit-learn==1.4.1.post1
nltk==3.8.1
plotly==5.18.0
chardet==5.2.0
openpyxl==3.1.2 